% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_median_vector.R
\name{multiply_median_vector}
\alias{multiply_median_vector}
\title{Multiply each element of a vector by the vector's median}
\usage{
multiply_median_vector(vec)
}
\arguments{
\item{vec}{A numeric vector of some length}
}
\value{
the resulting vector
}
\description{
Multiply each element of a vector by the vector's median
}
\examples{
multiply_median_vector(vec = c(1, 4, 5, 2, 9))
}
