% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonsense_multiply.R
\name{nonsense_multiply}
\alias{nonsense_multiply}
\title{A random system of equations}
\usage{
nonsense_multiply(x, y)
}
\arguments{
\item{x}{any number, unitless}

\item{y}{any number, unitless}
}
\value{
a number produced by plugging in x and y
}
\description{
A random system of equations
}
